Annotation update

+++ annotations/nocedit.exe/pseudocode/src/core/box.cpp/CBoundingBox3D_isVisible_FUN_004204f0.cpp
@@ -52,12 +52,12 @@ int __cdecl core_box_cpp_CBoundingBox3D_isVisible_FUN_004204f0(CBoundingBox3D *t
     local_20.y = local_20.y + local_10;
     local_20.z = local_20.z + local_c;
     if (g_CurrentSceneCamera == (CDemonCamera *)0x0) {
-      g_PerspectiveReciprocal = (float)g_CurrentSceneCamera;
+      g_PerspectiveReciprocal = (int)g_CurrentSceneCamera;
     }
     else {
       core_dcamera_cpp_CDemonCamera_setupPerspectiveAndFog_FUN_004537d0
-                (&g_CDemonCameraInstance,&local_20,(float)local_74);
-      if (65000 < (int)g_PerspectiveReciprocal) {
+                (&g_CDemonCameraInstance,&local_20,local_74);
+      if (65000 < g_PerspectiveReciprocal) {
         return 0;
       }
     }
